Release held requests sequentially
This makes it so that held requests are released sequentially, rather
than in parallel. This is important for our chat app as messages need
to arrive at the server in-order.

I understand if this isn't necessary for your app/use case, maybe this
should be an option instead?

> Note: I have 0 experience with observables, so just let me know else
to make this happen if there's a better way.